+++ fabric/src/main/java/dev/thenu/ReddensStoneLanternReconstructed/blockType/SmallproofStoneLanternLightBlock.java
@@ -1,6 +1,7 @@
 package dev.thenu.ReddensStoneLanternReconstructed.blockType;
 
 import dev.thenu.ReddensStoneLanternReconstructed.init.BlockFile;
+import dev.thenu.ReddensStoneLanternReconstructed.networking.clickProcedure.rightClick.SmallproofRightClickProcedure;
 import dev.thenu.ReddensStoneLanternReconstructed.world.blockPlaceProcedure.LanternPlaceSoundProcedure;
 import net.minecraft.block.*;
 import net.minecraft.entity.player.PlayerEntity;
@@ -76,7 +77,7 @@ public class SmallproofStoneLanternLightBlock extends Block implements Waterlogg
         boolean flag = context.getWorld().getFluidState(context.getBlockPos()).getFluid() == Fluids.WATER;
         return super.getPlacementState(context).with(WATERLOGGED, flag);
     }
-/*
+
     @Override
     public ItemStack getPickStack(WorldView world, BlockPos pos, BlockState state, boolean includeData) {
         return new ItemStack(BlockFile.SMALLPROOF_STONE_LANTERN_LIGHT);
@@ -96,12 +97,8 @@ public class SmallproofStoneLanternLightBlock extends Block implements Waterlogg
     @Override
     protected ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit) {
         if (!world.isClient()) {
-            BlockState newState = BlockFile.SMALLPROOF_STONE_LANTERN_DARK.getDefaultState()
-                    .with(WATERLOGGED, state.get(WATERLOGGED));
-
-            world.setBlockState(pos, newState, Block.NOTIFY_ALL);
-            world.playSound(null, pos, SoundEvents.ITEM_FLINTANDSTEEL_USE, net.minecraft.sound.SoundCategory.BLOCKS, 0.5F, 1.0F);
+            SmallproofRightClickProcedure.execute(world, pos.getX(), pos.getY(), pos.getZ());
         }
         return ActionResult.SUCCESS;
-    }*/
+    }
 }