+++ fabric/src/main/java/dev/thenu/ReddensStoneLanternReconstructed/blockType/ThinStoneLanternBlockDarkBlock.java
@@ -38,7 +38,7 @@ public class ThinStoneLanternBlockDarkBlock extends Block implements Waterloggab
         this.setDefaultState(this.stateManager.getDefaultState().with(WATERLOGGED, false));
     }
 
-
+/*
     @Override
     public VoxelShape getCameraCollisionShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
         return VoxelShapes.empty();
@@ -95,6 +95,6 @@ public class ThinStoneLanternBlockDarkBlock extends Block implements Waterloggab
     protected ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, BlockHitResult hit) {
         ThinStoneLanternBlockRightClickProcedure.execute(world, pos.getX(), pos.getX(), pos.getX());
         return ActionResult.SUCCESS;
-    }
+    }*/
 
 }